+++ db/migrate/20190911140550_create_clientes.rb
@@ -1,7 +1,7 @@
 class CreateClientes < ActiveRecord::Migration[5.2]
   def change
     create_table :clientes do |t|
-      t.integer :documento
+      t.bigint :documento
       t.string :nombres
       t.string :apellidos
       t.string :correo
